+++ timepiece/src/main/java/com/appcenter/timepiece/service/OAuth2Attributes.java
@@ -10,7 +10,7 @@ import java.util.Map;
 @Getter
 @ToString
 @Slf4j
-public class OAuthAttributes {
+public class OAuth2Attributes {
     private Map<String, Object> attributes;     // OAuth2 반환하는 유저 정보
     private String nameAttributesKey;
     private String name;
@@ -21,8 +21,8 @@ public class OAuthAttributes {
     private String provider;
 
     @Builder
-    public OAuthAttributes(Map<String, Object> attributes, String nameAttributesKey,
-                           String name, String email, String gender, String ageRange, String profileImageUrl, String provider) {
+    public OAuth2Attributes(Map<String, Object> attributes, String nameAttributesKey,
+                            String name, String email, String gender, String ageRange, String profileImageUrl, String provider) {
         this.attributes = attributes;
         this.nameAttributesKey = nameAttributesKey;
         this.name = name;
@@ -33,15 +33,15 @@ public class OAuthAttributes {
         this.provider = provider;
     }
 
-    public static OAuthAttributes of(String socialName, Map<String, Object> attributes) {
+    public static OAuth2Attributes of(String socialName, Map<String, Object> attributes) {
         if ("kakao".equals(socialName)) {
             return ofKakao("id", attributes);
         }
         return ofGoogle("sub", attributes);
     }
 
-    private static OAuthAttributes ofGoogle(String userNameAttributeName, Map<String, Object> attributes) {
-        return OAuthAttributes.builder()
+    private static OAuth2Attributes ofGoogle(String userNameAttributeName, Map<String, Object> attributes) {
+        return OAuth2Attributes.builder()
                 .name(String.valueOf(attributes.get("name")))
                 .email(String.valueOf(attributes.get("email")))
                 .profileImageUrl(String.valueOf(attributes.get("picture")))
@@ -51,11 +51,11 @@ public class OAuthAttributes {
                 .build();
     }
 
-    private static OAuthAttributes ofKakao(String userNameAttributeName, Map<String, Object> attributes) {
+    private static OAuth2Attributes ofKakao(String userNameAttributeName, Map<String, Object> attributes) {
         Map<String, Object> kakaoAccount = (Map<String, Object>) attributes.get("kakao_account");
         Map<String, Object> kakaoProfile = (Map<String, Object>) kakaoAccount.get("profile");
 
-        return OAuthAttributes.builder()
+        return OAuth2Attributes.builder()
                 .name(String.valueOf(kakaoProfile.get("nickname")))
                 .email(String.valueOf(kakaoAccount.get("email")))
                 .gender(String.valueOf(kakaoAccount.get("gender")))
